content_bot/requirements.txt
anthropic
requests
requests-oauthlib
moviepy
pillow
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
pytrends
feedparser
